backend/requirements.txt
# Django and DRF
Django==5.0.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.9

# Image Processing
Pillow==10.2.0
rembg==2.0.53

# Asynchronous Tasks
celery==5.3.6
redis==5.0.3

# Environment and WSGI
python-dotenv==1.0.1
gunicorn==21.2.0
